+++ .gitignore
@@ -3,3 +3,8 @@
 /*.log
 /lib/
 /build/
+
+/dist
+/dist.browser
+
+/example/index.esm.js
