+++ .yarnrc
@@ -1,0 +1,3 @@
+version-tag-prefix ""
+version-git-message "chore: bump up to %s"
+
