+++ package.json
@@ -1,13 +1,22 @@
 {
   "name": "request-animation-frame-polyfill",
-  "main": "lib/request-animation-frame-polyfill.js",
   "version": "1.0.2",
   "description": "requestAnimationFrame polyfill library",
+  "main": "dist/request-animation-frame-polyfill.umd.js",
+  "types": "dist/request-animation-frame-polyfill.d.ts",
+  "files": [
+    "dist"
+  ],
+  "module": "dist/request-animation-frame-polyfill.esm.min.js",
   "scripts": {
     "clean": "rimraf lib && rimraf build",
-    "build": "mkdirp lib && babel src/index.ts > ./lib/request-animation-frame-polyfill.js && npx tsc && mv build/index.d.ts ./lib/request-animation-frame-polyfill.d.ts",
-    "test": "babel-node --extensions \".ts\" -- ./src/index.test.ts",
-    "prepare": "yarn run clean && yarn build && yarn run test"
+    "test": "jest --coverage",
+    "build:types": "tsc -t esnext --moduleResolution node -d --emitDeclarationOnly --outFile dist/request-animation-frame-polyfill.d.ts src/index.ts",
+    "build:js": "rollup -c rollup.config.js",
+    "build:minjs:umd": "terser dist/request-animation-frame-polyfill.umd.js --compress --mangle > dist/request-animation-frame-polyfill.umd.min.js",
+    "build:minjs:esm": "terser dist/request-animation-frame-polyfill.esm.js --compress --mangle > dist/request-animation-frame-polyfill.esm.min.js",
+    "build:minjs": "npm run build:minjs:esm -s && npm run build:minjs:umd -s",
+    "build": "npm run build:js -s && npm run build:minjs -s && npm run build:types -s"
   },
   "repository": {
     "type": "git",
@@ -26,16 +35,18 @@
     "url": "https://github.com/aluc-io/request-animation-frame-polyfill/issues"
   },
   "devDependencies": {
-    "@babel/cli": "^7.6.4",
-    "@babel/core": "^7.6.4",
-    "@babel/node": "^7.6.3",
-    "@babel/plugin-transform-runtime": "^7.6.2",
-    "@babel/preset-env": "^7.6.3",
-    "@babel/preset-typescript": "^7.6.0",
-    "@types/node": "^12.12.5",
-    "mkdirp": "^0.5.1",
+    "@rollup/plugin-typescript": "^8.1.1",
+    "@types/jest": "^26.0.20",
+    "@types/node": "12.19.15",
+    "jest": "^26.6.3",
     "rimraf": "^3.0.0",
-    "typescript": "^3.6.4"
+    "rollup": "^2.38.5",
+    "terser": "^5.6.0-beta",
+    "ts-jest": "^26.5.0",
+    "ts-node": "^9.1.1",
+    "tslib": "^2.1.0",
+    "typedoc": "^0.20.21",
+    "typescript": "^4.1.3"
   },
   "dependencies": {}
 }
